Modify : 회원가입, 로그인 시 Refresh Token 발행

+++ back_Nanal/src/main/java/com/dbd/nanal/repository/UserProfileRepository.java
@@ -2,11 +2,12 @@ package com.dbd.nanal.repository;
 
 import com.dbd.nanal.model.UserEntity;
 import com.dbd.nanal.model.UserProfileEntity;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserProfileRepository extends JpaRepository<UserProfileEntity, Long> {
+public interface UserProfileRepository extends JpaRepository<UserProfileEntity, Integer> {
 
     UserProfileEntity findByUser(UserEntity user);
     UserProfileEntity findByNickname(String nickname);
